fix: error on CSSLint stderr
If CSSLint returns text on `stderr`, or more likely a message is thrown
for a miss-configured path to CSSLint, show this message to the user
instead of attempting to go on and parse the (empty) output.

Messages are shown until the user interacts with them, but are
deduplicated so they are only actively shown once.

Fixes #181.
Fixes #189.